refactor (search): move classes into sparql package
