+++ webapi/src/main/scala/org/knora/webapi/responders/v2/ResourcesResponderV2.scala
@@ -42,7 +42,7 @@ import org.knora.webapi.util.ConstructResponseUtilV2.{MappingAndXSLTransformatio
 import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util._
 import org.knora.webapi.util.date.CalendarNameGregorian
-import org.knora.webapi.util.search.ConstructQuery
+import org.knora.webapi.util.search.sparql.ConstructQuery
 import org.knora.webapi.util.search.sparql.gravsearch.GravsearchParser
 
 import scala.concurrent.Future
