+++ webapi/src/main/scala/org/knora/webapi/responders/v2/SearchResponderV2.scala
@@ -28,8 +28,8 @@ import org.knora.webapi.messages.v2.responder.searchmessages._
 import org.knora.webapi.util.ActorUtil._
 import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util._
-import org.knora.webapi.util.search.ApacheLuceneSupport.{CombineSearchTerms, MatchStringWhileTyping}
-import org.knora.webapi.util.search._
+import org.knora.webapi.util.search.sparql.ApacheLuceneSupport.{CombineSearchTerms, MatchStringWhileTyping}
+import org.knora.webapi.util.search.sparql._
 import org.knora.webapi.util.search.sparql.gravsearch.GravsearchUtilV2.SparqlTransformation._
 import org.knora.webapi.util.search.sparql.gravsearch._
