+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/ApacheLuceneSupport.scala
@@ -1,4 +1,4 @@
-package org.knora.webapi.util.search
+package org.knora.webapi.util.search.sparql
 
 
 /**
