+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/QueryTraverser.scala
@@ -17,7 +17,7 @@
  * License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.knora.webapi.util.search
+package org.knora.webapi.util.search.sparql
 
 import org.knora.webapi.util.search.sparql.gravsearch.GravsearchTypeInspectionResult
