+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/SparqlQuery.scala
@@ -17,7 +17,7 @@
  * License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.knora.webapi.util.search
+package org.knora.webapi.util.search.sparql
 
 import akka.http.scaladsl.model.{HttpCharsets, MediaType}
 import org.knora.webapi._
