+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/gravsearch/AnnotationReadingGravsearchTypeInspector.scala
@@ -23,7 +23,7 @@ import akka.actor.ActorSystem
 import org.knora.webapi._
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
 import org.knora.webapi.util.SmartIri
-import org.knora.webapi.util.search._
+import org.knora.webapi.util.search.sparql._
 import org.knora.webapi.util.search.sparql.gravsearch.GravsearchTypeInspectionUtil.{TypeAnnotationProperties, TypeAnnotationProperty}
 
 import scala.concurrent.Future
