+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/gravsearch/GravsearchTypeInspectionRunner.scala
@@ -21,7 +21,7 @@ package org.knora.webapi.util.search.sparql.gravsearch
 
 import akka.actor.ActorSystem
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
-import org.knora.webapi.util.search._
+import org.knora.webapi.util.search.sparql._
 import org.knora.webapi.{GravsearchException, KnoraDispatchers, OntologyConstants}
 
 import scala.concurrent.{ExecutionContext, Future}
