+++ webapi/src/main/scala/org/knora/webapi/util/search/sparql/gravsearch/GravsearchUtilV2.scala
@@ -23,7 +23,7 @@ import org.knora.webapi._
 import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
 import org.knora.webapi.messages.store.triplestoremessages.{SparqlConstructResponse, SparqlSelectResponse, VariableResultsRow}
 import org.knora.webapi.util.{ConstructResponseUtilV2, ErrorHandlingMap, SmartIri, StringFormatter}
-import org.knora.webapi.util.search._
+import org.knora.webapi.util.search.sparql._
 import org.knora.webapi.util.IriConversions._
