+++ webapi/src/main/twirl/queries/sparql/v2/searchFulltext.scala.txt
@@ -19,7 +19,7 @@
 
 @import org.knora.webapi.IRI
 @import org.knora.webapi.SparqlGenerationException
-@import org.knora.webapi.util.search.ApacheLuceneSupport.CombineSearchTerms
+@import org.knora.webapi.util.search.sparql.ApacheLuceneSupport.CombineSearchTerms
 
 @*
  * Performs a simple full-text search.
