+++ webapi/src/main/twirl/queries/sparql/v2/searchResourceByLabelSubQuerySelectGraphDB.scala.txt
@@ -19,7 +19,7 @@
 
 @import org.knora.webapi.IRI
 @import org.knora.webapi.SparqlGenerationException
-@import org.knora.webapi.util.search.ApacheLuceneSupport.MatchStringWhileTyping
+@import org.knora.webapi.util.search.sparql.ApacheLuceneSupport.MatchStringWhileTyping
 
 @*
  * Select subquery for a search for resources by their label.
