+++ webapi/src/test/scala/org/knora/webapi/responders/v2/SearchResponderV2SpecFullData.scala
@@ -9,7 +9,7 @@ import org.knora.webapi.messages.v2.responder.valuemessages._
 import org.knora.webapi.twirl.{StandoffTagIriAttributeV2, StandoffTagV2}
 import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util.StringFormatter
-import org.knora.webapi.util.search._
+import org.knora.webapi.util.search.sparql._
 import org.knora.webapi.{ApiV2Simple, InternalSchema, Settings}
 
 class SearchResponderV2SpecFullData(implicit stringFormatter: StringFormatter) {
